# openapi.yaml
openapi: 3.1.0
info:
  title: Snake Arena Online Backend API
  version: 1.0.0
  description: >
    HTTP API expected by the frontend BackendService in frontend/src/services.
    Authenticated endpoints use the bearer token returned by login or signup.
servers:
  - url: /api
security: []
...
paths:
  /auth/signup:
    post:
      tags: [Auth]
      summary: Create an account and start a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "200":
          description: Account created and authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Username already taken.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate an existing user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "200":
          description: Authenticated user and bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      tags: [Auth]
      summary: End the current session
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Logged out.
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: jwt
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication is required or the token is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    User:
      type: object
      additionalProperties: false
      required: [id, username]
      properties:
        id:
          type: string
          example: u_ab12cd34
        username:
          type: string
          example: demo
    Credentials:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          example: demo
        password:
          type: string
          minLength: 1
          format: password
          example: demo
    AuthResult:
      type: object
      additionalProperties: false
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
          description: Bearer token used for authenticated endpoints.
          example: mock
...
    Error:
      type: object
      additionalProperties: false
      required: [message]
      properties:
        message:
          type: string
          example: Not authenticated